+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##0.1.5 (-- --- 2013
 - Git updater failed (https://github.com/PierreRambaud/GotCms/issues/24)
 - Add translation directory for installation
+- Add captcha for blog module (https://github.com/PierreRambaud/GotCms/issues/25)
 
 ## 0.1.4 (12 Feb 2013):
 - Bug with HTTP 404 (undefined variable view)